chore: Update docker-compose.yml for nrss service
This commit updates the docker-compose.yml file to rename the service from "node_new" to "nrss". It also updates the image tag to "isvaku/notion-rss:v1.0.1". Additionally, it modifies the cronTime for the "getEntriesJob" to run every hour and the "createNotionPagesJob" to run every 10 minutes.

These changes ensure that the nrss service is properly configured and aligned with the latest requirements.