chore: Update build-docker-image and push-docker-image scripts
This commit updates the build-docker-image and push-docker-image scripts in the package.json file. It adds the -t flag to the build-docker-image script to tag the image with both the version number and "latest". It also modifies the push-docker-image script to push all tags of the image. These changes ensure that the Docker image is properly tagged and pushed to the repository.